Introduce new reader classes and factory to simplify their management
Split the long vtkF3DMetaReader class in multiple reader classes.
This allows better management of the reader and introspection. Two new options are introduced: --extensions-list and --readers-list.
F3DReader is the basic class for every file format readers. It must provide information about its name and descriptions, the managed file formats, and must be able to produce either a VTK reader or a VTK importer.
Each F3DReader need to be added to the F3DReaderFactory singleton. To make it simple and automatic, you just have to declare them with a CMake macro. CMake will generate the class F3DReaderInstantiator which automatically instantiates all the reader classes and registers them to the factory. One instance of this class is hold by F3DLoader.