Rework of config and option management in F3DStarter in order to improve the behavior.
 - add a "global" category in the config file that as a different behavior as other category
 - rework the option logic

Previous behavior:

When loading a file, the command line and config file are parsed in order to figure out which options to use

New behavior:

F3D has a dynamic option that will changed over time when interacting with F3D.
When loading a file for the first time, F3D keep a hand on:

 - A dynamic option that will evolve overtime
 - Used options with the file config applied to it and copied to the engine
 - The option in use in the engine (hold by engine)

If no file was load previously, dynamic option contains the command line option + global section of config file.
file options is copied from dynamic option and config for the file being loaded is added on it, and command line option are applied on top of it.
file options is then copied to the engine to be used

If a file was loaded previously, before doing anything else, F3D will compare each option between the current engine options and the previous file options.
Each different option is stored in the dynamic option.
File options is copied from dynamic option and config for the file being loaded is added on it.
Command line options is NOT added on top of it.
File options is then copied to the engine to be used.