refactor: Make do_reverse_rotate default to true
I am keeping this available for now in anticipation of
edge cases where we do need to perform the operations.

When the code is use with my production assets, we might
encounter such edge cases.

When the follow release of f3d comes along and we still
have not encounter such edge cases, I can remove it.

Signed-off-by: Nicholas Yue <yue.nicholas@gmail.com>