assimp: add data validation flag
Passing this flag makes assimp perform validation of file before post-processing
which can give a chance for assimp to report an error of some kind.
In particular case when issue was detected FBX file was problematic
to assimp and without check it could lead to crash.

Related test case is also added.