add `dir` multi-file mode (#2193)
Add dir multi-file mode to group files according to their parent directory.

I refactored the grouping logic to use a group key computed from the file path according to the mode instead of having separate logic for each.

The test case is kind of ugly and has a warning but it does hit the recursive case for file addition with existing data. It can be improved when I add a proper dataset to test regex grouping in a future PR.