CI: Add logic to trigger f3d-docker-image dispatch workflow (#2202)
f3d-docker-images needs to be manually dispatched when changing the timestamp.
This is cumbersome and require manual operation, this automatize this step.

A similar design may allow to enable VTK nightly builds again for android and wasm

Please note this REQUIRES a read/write PAT in a secract named: F3D_DOCKER_CI_DISPATCH to be able to run the action.